feat: improve the responsiveness of the shape change
There was a "lag" in the shape update because we did not hook the animation_changed signal.
Now we do, and it's waaay smoother.

We also added an "update" mode that does not replace the whole shape,
but only updates its properties, and therefore does not trigger extra "entered" signals.

Fixes #5

/spend 2h30