feat: add the INTERPOLATE shape update mode
This helps a lot with physics continuity.
I don't see any clipping now, when using this mode with my solidboxes.
This works even better if you set your physics engine to 120 FPS or more.

ADR
---

Tweening on the properties of the shapes is quite verbose,
but it's very simple code and it's faster than using reflection.

I had to make a call (maintainability vs performance),
and I chose performance because the maintenance overhead is not that big.

---

Fixes #7

/spend 4h30